requirements.txt
about-time==4.2.1
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
av==14.0.1
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clipsai==0.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
ctranslate2==4.4.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
facenet-pytorch==2.6.0
facexlib==0.3.0
faster-whisper==1.1.0
fastjsonschema==2.21.1
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gfpgan==1.3.8
grapheme==0.6.0
grpcio==1.69.0
huggingface-hub==0.27.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
ipython==8.12.3
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
lmdb==1.6.2
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.20
mistune==3.1.0
ml_dtypes==0.5.1
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-ml-py==12.560.30
omegaconf==2.3.0
onnxruntime==1.20.1
openai-whisper==20240930
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
primePy==1.3
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
realesrgan==0.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.1
scenedetect==0.6.5.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
semver==3.0.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tb-nightly==2.19.0a20250109
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.2.2
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==1.6.1
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whisperx==3.3.0
yapf==0.43.0
yarg==0.1.9
yarl==1.18.3
